+++ easyadapter/RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## 1.1.0-alpha
 ##### Easyadapter
 * ANDDEP-646 - Improved EasyAdapter. In BaseItem added next and previous links to items, adapterPosition, position in ItemList.
+* ANDDEP-795 Added option in EasyAdapter for custom set SpanSizeLookup when GridLayoutManager attached; not changing by default
 ## 1.0.0
 ##### Easyadapter
 * ANDDEP-270 Added support for asynchronous view inflate in ViewHolder